initial reorganization
