+++ src/lib/maintenance/generate_mapping/writer.py
