+++ tests/test_mapping_template_generator.py
@@ -5,7 +5,7 @@ import tempfile
 import unittest
 from pathlib import Path
 
-from src.lib.generate_mapping import flatten_schema, write_mapping_template_csv
+from src.lib.maintenance.generate_mapping import flatten_schema, write_mapping_template_csv
 
 ROOT = Path(__file__).resolve().parents[1]
 FIXTURES = ROOT / "tests" / "fixtures"
